navigation bar element added
